refactor(cli): extract lightweight output helpers; add collections delete --json
Address PR review feedback:

- Move the pure table/JSON/prompt helpers out of deploy-helpers.ts (which
  pulls in the heavy deploy stack at top level) into a dependency-free
  output.ts. deploy-helpers re-exports them for existing callers; the
  lightweight command groups (collections, costs, models-hf,
  models-recommended, worker-models) now import from output.ts so they no
  longer load deploy-only transitive deps at startup.
- Give `collections delete` a --json flag for consistency with the rest of
  the group.
